// src/app/shared/interfaces/user.ts
export interface User {
  img: string;
  uid: string;
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  isAdmin: boolean;
  skills: any;
  country: string;
  city: string;
  university: string;
  speciality: string;
  yearOfAdmission: string;
  yearOfEnding: string;
  isOnline: boolean;
}